Enabled HACS support
