+++ custom_components/grafik_eye_3000/manifest.json
@@ -1,0 +1,10 @@
+{
+  "domain": "grafik_eye",
+  "name": "Lutron Grafik Eye 3000",
+  "codeowners": ["@nigelellis"],
+  "documentation": "https://github.com/nigelellis/grafik_eye_3000",
+  "issue_tracker": "https://github.com/nigelellis/grafik_eye_3000/issues",
+  "loggers": ["grafik_eye"],
+  "requirements": [],
+  "version": []
+}
